Added requested changes - TestHelper.swift, changed the forced unwraping
